Re-structured app with Blueprints
