+++ api/app.py
@@ -1,8 +1,6 @@
 # Entry point for flask app
 import datetime
 from flask import Flask
-from flask_restful import Api
-from api.__init__ import app
 from flask_jwt import JWT, jwt_required, current_identity
 from api.resources.v1.meals import Meal, MealList
 from api.resources.v1.orders import Order, OrderList
@@ -11,10 +9,8 @@ from api.resources.v1.users import User
 from api.resources.v1.registration import Registration
 from api.resources.v1.login import Login
 
-flask_app = app
 
-def setup_routes():
-    api = Api(flask_app) 
+def setup_routes(api):
 
     api.add_resource(Registration,'/api/v1/auth/register/')
     api.add_resource(Login, '/api/v1/auth/login/')
