+++ tests/test_auth_login.py
@@ -1,27 +1,20 @@
-import unittest
 import json
-from api.app import app
+from tests.test_main import MainTest
 
-class LoginTestCase(unittest.TestCase):   
-    ''' Tests for user login authentication '''
-    def setUp(self):
-        self.app = app
-        self.app.testing = True
-        self.client = self.app.test_client
-
-        
+class LoginTestCase(MainTest):   
+    ''' Tests for user login authentication '''      
 
     '''Test if user logged in successfully'''
     def test_registered_user_log_in(self):
         test_data = {"first_name": "Joshua", "last_name": "Mugisha", "user_name": "jmugisha", "email": "joshua@example.com", "password": "@password"}
-        res = self.client().post('/api/v1/auth/register/', data=json.dumps(test_data))
+        res = self.client.post('/api/v1/auth/register/', data=json.dumps(test_data))
         registered_data = json.loads(res.data.decode())
         self.assertTrue(
             registered_data['message'] == 'User {} was created'.format(test_data['user_name'])
         )
         self.assertEqual(res.status_code, 201)
 
-        res1 = self.client().post('/api/v1/auth/login/', data=json.dumps({"email": "joshua@example.com", "password": "@password"}),content_type='application/json')
+        res1 = self.client.post('/api/v1/auth/login/', data=json.dumps({"email": "joshua@example.com", "password": "@password"}),content_type='application/json')
         data = json.loads(res1.data.decode())
         self.assertTrue(data['message'] == 'Successfully logged in')
         self.assertTrue(data['access_token'])
@@ -31,13 +24,13 @@ class LoginTestCase(unittest.TestCase):
     '''Test if a user attempted to login in with missing credentials'''
     def test_user_logging_in_with_missing_credentials(self):
         test_data = json.dumps({'username': '','email': '', 'password': ''}) 
-        res = self.client().post('/api/v1/auth/login/', data=test_data, content_type='application/json')
+        res = self.client.post('/api/v1/auth/login/', data=test_data, content_type='application/json')
         result = json.loads(res.data.decode())
         self.assertIn('Bad Request',result['error']) 
         self.assertEqual(res.status_code, 401)
 
     def test_unregistered_user_login(self):
-        response = self.client().post(
+        response = self.client.post(
             '/api/v1/auth/login/',
             data=json.dumps(dict(
                 email='katarikawe@gmail.com',
